In progress for improving the audio module
Going over all the current core feature and improving them.

+++ crates/mani/fabled_audio/src/container/raw_clip.rs
@@ -1,12 +1,13 @@
 // You shouldn't have to directly call method from both Raw struct.
+// todo should the audio effect start here?
 
-pub struct RawClip<T: rodio::Source<Item = i16>> {
+pub struct RawClip<T: rodio::Source<Item = f32>> {
     data: T,
 }
 
 impl<T> AsRef<T> for RawClip<T>
 where
-    T: rodio::Source<Item = i16>,
+    T: rodio::Source<Item = f32>,
 {
     fn as_ref(&self) -> &T {
         &self.data
@@ -15,7 +16,7 @@ where
 
 impl<T> RawClip<T>
 where
-    T: rodio::Source<Item = i16>,
+    T: rodio::Source<Item = f32>,
 {
     pub fn new(data: T) -> Self {
         Self { data }
